# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FDC+ Serial Drive Server API",
    "version": "3.0.0-alpha",
    "description": "REST API for the Altair 8800 FDC+ Serial Drive Server web interface. Manages disk image serving, CP/M filesystem browsing, cassette audio, terminal serial connections, script replay, and XMODEM transfers.\n\n## Socket.IO Events\n\nThe server also exposes real-time events via Socket.IO (not covered by OpenAPI).\n\n### Server → Client\n\n| Event | Payload | Description |\n|-------|---------|-------------|\n| `status` | `ServerStatus` | Periodic server status broadcast (1s interval) |\n| `terminal:data` | `number[]` | Incoming serial data as byte array |\n| `terminal:status` | `TerminalStatus` | Terminal connection status change |\n| `terminal:error` | `{message: string}` | Terminal error |\n| `replay:progress` | `ReplayProgress` | Transfer progress update |\n| `replay:status` | `{active, mode, progress}` | Transfer status on connect |\n\n### Client → Server\n\n| Event | Payload | Description |\n|-------|---------|-------------|\n| `request-status` | _(none)_ | Request current server status |\n| `terminal:write` | `string` | Send keyboard input to terminal serial port |\n| `terminal:control` | `{type, value}` | Set DTR/RTS control signals |\n| `replay:start` | `ReplayStartParams` | Start raw replay or XMODEM send |\n| `replay:cancel` | _(none)_ | Cancel active transfer |"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "components": {
    "schemas": {
...
      "InstanceCreateRequest": {
        "type": "object",
        "description": "Create a Machine Instance from a stored Profile ref, a preset id, or an inline MachineProfile (exactly one).",
        "properties": {
          "profileRef": {
            "type": "string",
            "example": "my-imsai@1.0.0",
            "description": "Stored Machine Profile reference (name@version, or a bare name → latest)."
          },
          "preset": {
            "type": "string",
            "example": "imsai-cpm",
            "description": "Built-in preset id (see GET /api/instances/presets)."
          },
          "profile": {
            "type": "object",
            "additionalProperties": true,
            "description": "Inline MachineProfile (cpuKind, clock, resetVector, memory[], cards[], consoleCardId)."
          },
          "speed": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "enum": [
                  "max"
                ]
              }
            ],
            "description": "Launch speed: Hz (e.g. 2000000 for authentic 2 MHz) or \"max\"."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/instances": {
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "List virtual Machine Instances",
        "responses": {
          "200": {
            "description": "Array of instances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instances": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MachineInstance"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Define a persistent Machine Instance (not started)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The defined instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/transient": {
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Create and start a transient Machine Instance",
        "description": "Memory-only; boots immediately and leaves no residue on destroy. Mount a boot disk first.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The running instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{id}": {
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "Get a Machine Instance by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Instances"
        ],
        "summary": "Destroy a Machine Instance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destroyed"
          }
        }
      }
    },
    "/api/instances/{id}/start": {
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Start (or resume) a Machine Instance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The running instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{id}/stop": {
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Stop a running Machine Instance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stopped instance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{id}/speed": {
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Change a running instance's speed live (no restart)",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "speed"
                ],
                "properties": {
                  "speed": {
                    "oneOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string",
                        "enum": [
                          "max"
                        ]
                      }
                    ],
                    "description": "Hz (e.g. 2000000, 4000000) or 'max'."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The instance at its new speed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instance": {
                      "$ref": "#/components/schemas/MachineInstance"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Instance is not running"
          }
        }
      }
    },
    "/api/instances/{id}/snapshots": {
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "List an instance's disk/media snapshots",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshots, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshots": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InstanceSnapshot"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Snapshot an instance's disk/media state (FR-18)",
        "description": "Captures the machine definition + each bound drive's disk state (execution state is out of scope).",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "snapshot": {
                      "$ref": "#/components/schemas/InstanceSnapshot"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/instance-snapshots/{snapshotId}/restore": {
      "post": {
        "tags": [
          "Instances"
        ],
        "summary": "Restore a disk/media snapshot onto its instance",
        "description": "Stops the instance, writes the captured disks back, and restarts it (reproduces the disk state; the machine reboots).",
        "parameters": [
          {
            "in": "path",
            "name": "snapshotId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restored"
          },
          "404": {
            "description": "Snapshot or instance not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Instances"
        ],
        "summary": "Delete a disk/media snapshot",
        "parameters": [
          {
            "in": "path",
            "name": "snapshotId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  }
}